% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaca.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate test-statistic covariance.}
\usage{
estimate_sigma(expression_data, trt)
}
\arguments{
\item{expression_data}{the expression matrix.}

\item{trt}{sample labels. 0 for control and 1 for treatment}
}
\value{
a list 
\item{sigma}{estimated covariance matrix for the gene-level test statistics. }
\item{t_val}{the gene level statistics, one for each gene}
}
\description{
Estimate test-statistic covariance
}

