% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaca.R
\name{meaca_multiple}
\alias{meaca_multiple}
\title{meaca-multiple.}
\usage{
meaca_multiple(expression_data, trt, geneset, standardize = T,
  minSetSize = 5, fdr_method = "BH")
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment labels, 1 for treatment 0 for control.}

\item{geneset}{gene sets to be tested, an object from \code{read_gene_set}.}

\item{standardize}{whether the data should be standaridzed.}

\item{minSetSize}{the minimum number of genes contained in a gene set for it to be considered as a test set.}

\item{fdr_method}{which method is ued to adjust the p values. Options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", see arguments in function \code{p.adjust} for more details.}
}
\value{
a data frame with the following columns
\item{set_name}{the name of the gene set }
\item{set_sizel}{the size of the gene set}
\item{status}{"Up" for up-regulated gene sets and "Down" for down-regulated gene sets} 
\item{p1}{chi square test p value}
\item{p1_fdr}{adjusted p1, using \code{p.adjust} with appropriate method}
\item{p2}{a two-sided t-test p value equivalent to the chi square test}
\item{p2_fdr}{adjusted p2 }
}
\description{
meaca for testing multiple gene sets.
}
\examples{
m <- 100   # number of rows (genes)
n <- 50    # number of columns (samples)
y <- matrix(rnorm(5000), 100, 50)              # expression  matrix
rownames(y) <- paste("gene", 1:m, sep = "")  
colnames(y) <- paste("sample", 1:n, sep = "")
trt <- rep(c(0, 1), each = n/2)                # treatment labels
## create the gene set format, desired for this function.
total <- 15                     # number of gene sets in the database
set.seed(100)
size <- sample(1:30, replace=F, total)
go_term <- rep(0, m)
gs_line <- list()
set_name <- paste("set", 1:length(size), sep = "")
for (i in 1:total) {
    set.seed(i)
    gs_line[[i]] <- rownames(y)[sample(1:m, size[i])] 
}
gs <- list(total=total, size=size,set_name = set_name, gene_set=gs_line)
# run the multiple gene set test
result <- meaca_multiple(y, trt, gs)
}

