% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaca.R
\name{meaca_single}
\alias{meaca_single}
\title{Testing the enrichment status of a single pre-defined gene set.}
\usage{
meaca_single(expression_data, trt, go_term, standardize = T)
}
\arguments{
\item{expression_data}{the expressoin matrix.}

\item{trt}{treatment labels, should be 0s and 1s.}

\item{go_term}{an indicator vector with value 1 for genes in the test, 0 otherwise.}

\item{standardize}{whether the data should be standardized. Set to be TRUE for real data analysis.}
}
\value{
a list 
\item{stat}{test statistic for meaca}
\item{p1}{chi square test p value}
\item{status}{"Up" for up-regulated gene sets and "Down" for down-regulated gene sets}
\item{p2}{a two-sided p value for conducting t-test from the test statstic \code{stat}}
}
\description{
meaca for single gene set test.
}
\examples{
 
m <- 100   # number of rows (genes)
n <- 50   # number of columns (samples)
y <- matrix(rnorm(5000), 100, 50)   # expression  matrix
trt <- rep(c(0, 1), each = n/2)      # treatment labels
go_term <- rep(0, m)
go_term[sample(1:m, 20)] <- 1     # the rows in the test set are labeled as 1
result <- meaca_single(y, trt, go_term)
}

